Check tasty file hash
