+++ compiler/src/dotty/tools/backend/jvm/GenBCode.scala
@@ -26,6 +26,7 @@ import scala.tools.asm
 import scala.tools.asm.tree._
 import tpd._
 import StdNames._
+import dotty.tools.dotc.core.tasty.TastyPickler
 import dotty.tools.io._
 
 class GenBCode extends Phase {
@@ -221,9 +222,9 @@ class GenBCodePipeline(val entryPoints: List[Symbol], val int: DottyBackendInter
                 val outstream = new DataOutputStream(outTastyFile.bufferedOutput)
                 try outstream.write(binary)
                 finally outstream.close()
-                // TASTY attribute is created but 0 bytes are stored in it.
-                // A TASTY attribute has length 0 if and only if the .tasty file exists.
-                Array.empty[Byte]
+                // TASTY attribute is created but only the header bytes are stored in it.
+                // A TASTY attribute has length `headerSize` if and only if the .tasty file exists.
+                java.util.Arrays.copyOf(binary, TastyPickler.headerSize)
               } else {
                 // Create an empty file to signal that a tasty section exist in the corresponding .class
                 // This is much cheaper and simpler to check than doing classfile parsing
